chore: refactor the component
