+++ web/src/types/resourceItem.d.ts
@@ -1,0 +1,11 @@
+interface ResourceProps {
+  resource: Resource;
+  handlecheckClick: () => void;
+  handleUncheckClick: () => void;
+  handleRenameBtnClick: (resource: Resource) => void;
+  handleDeleteResourceBtnClick: (resource: Resource) => void;
+  handlePreviewBtnClick: (resource: Resource) => void;
+  handleCopyResourceLinkBtnClick: (resource: Resource) => void;
+}
+
+type ResourceType = ResourceProps;
